% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{createWindow}
\alias{createWindow}
\title{Create new graphics window}
\usage{
createWindow(x, y, name, fullScreen = FALSE)
}
\arguments{
\item{x}{desired window width}

\item{y}{desired window height}

\item{name}{desired window name}

\item{fullScreen}{whether or not window is full screen}
}
\description{
Create new graphics window
}
